#21
добавил в extranceToCinema функцию выбора места
глобальна переменная всех дней
перенёс некоторые функции в заголовочный файл
