+++ cinema/AllMovDay_1.txt
@@ -1,17 +1,29 @@
-Кинозал 1                    День 1 
-- 10:00 - 11:30 (Шрек и болото багов)
-- 13:00 -14:30 (Дебаггер...)
-- 16:00 - 18:10 (Властелин лаб...)  
-- 20:00 - 22:00  (Гарри Кодер...)
+Кинозал 1                    День 1  
+1. 10:00 - 11:30  
+   Шрек и болото багов  
+2. 13:00 - 15:15  
+   Дебаггер 2: Судный день компиляции  
+3. 16:00 - 19:20  
+   Властелин лаб: Возвращение сеньора  
+4. 20:00 - 22:30  
+   Гарри Кодер и С++  
 
-Кинозал 2
-- 10:00 — 11:50 (Матрица данных)
-- 13:00 -14:45 (Унесенные сессией)
-- 16:00 - 18:00 (Титаник: крах программы) 
-- 20:00 - 22:10 (Властелин лаб...)  
+Кинозал 2  
+5. 10:00 — 12:15  
+   Матрица данных  
+6. 13:00 - 15:00  
+   Унесённые сессией  
+7. 16:00 - 19:15  
+   Титаник: крах программы  
+8. 20:00 - 23:20  
+   Властелин лаб: Возвращение сеньора  
 
-Кинозал 3 
-- 11:00 - 12:45 (Пятый алгоритм)
-- 14:00 - 15:45  (Люди в ч п)
-- 17:00 - 18:50 (Матрица данных) 
-- 20:00 - 22:10 (А дедлайны здесь жесткие)
+Кинозал 3  
+9. 11:00 - 13:00  
+   Пятый алгоритм  
+10. 14:00 - 15:30  
+    Люди в черных пикселях  
+11. 17:00 - 19:15  
+    Матрица данных  
+12. 20:00 - 23:10  
+    А дедлайны здесь жесткие
