+++ cinema/AllMovDay_2.txt
@@ -1,16 +1,27 @@
 Кинозал 1                    День 2  
-- 10:00 - 11:45  (Пятый алгоритм)
-- 13:00 — 15:00 (Титаник: крах программы) 
-- 16:00 - 18:10 (А дедлайны здесь жесткие)
-- 19:00 — 21:10  (Властелин лаб...)   
+1. 10:00 - 12:00  
+   Пятый алгоритм  
+2. 13:00 - 16:15  
+   Титаник: крах программы  
+3. 16:30 - 19:40  
+   А дедлайны здесь жесткие  
+4. 20:00 - 23:20  
+   Властелин лаб: Возвращение сеньора  
 
-Кинозал 2
-- 12:00 - 13:30 (Шрек и болото багов)
-- 16:00 - 18:00 (Назад к компилятору)
-- 20:00 - 21:30 (Дебаггер...)  
+Кинозал 2  
+5. 10:00 — 13:30  
+   Шрек и болото багов  
+6. 16:00 - 18:00  
+   Назад к компилятору  
+7. 20:00 - 22:15  
+   Дебаггер 2: Судный день компиляции  
 
-Кинозал 3
-- 11:00 - 12:50 (Матрица данных) 
-- 14:00 - 15:45 (Унесенные сессией)
-- 17:00 - 19:00 (Гарри Кодер) 
-- 20:00 - 22:15  (Лабные войны)
+Кинозал 3  
+8. 11:00 - 13:15  
+   Матрица данных  
+9. 14:00 - 16:00  
+   Унесённые сессией  
+10. 17:00 - 19:30  
+    Гарри Кодер  
+11. 20:00 - 22:15  
+    Лабные войны
