+++ cinema/AllMovDay_3.txt
@@ -1,16 +1,27 @@
 Кинозал 1                    День 3  
-- 11:00 - 12:45 (Люди в ч п)
-- 14:00 - 16:00  (Назад к компилятору)
-- 17:00 - 19:10 (А дедлайны здесь жесткие)  
-- 20:00 - 22:15  (Лабные войны)   
+1. 10:00 - 12:30  
+   Люди в черных пикселях  
+2. 14:00 - 16:00  
+   Назад к компилятору  
+3. 17:00 - 20:10  
+   А дедлайны здесь жесткие  
+4. 20:20 - 22:35  
+   Лабные войны  
 
-Кинозал 2
-- 10:00 - 11:30 (Дебаггер...)
-- 13:00 -14:30 (Шрек и болото багов)
-- 16:00 - 18:00  (Титаник: крах программы) 
-- 19:00 - 21:00 (Гарри Кодер)
+Кинозал 2  
+5. 10:00 — 12:15  
+   Дебаггер 2: Судный день компиляции  
+6. 13:00 - 14:30  
+   Шрек и болото багов  
+7. 16:00 - 19:15  
+   Титаник: крах программы  
+8. 19:30 - 22:00  
+   Гарри Кодер и С++  
 
-Кинозал 3
-- 12:00 -13:45 (Люди в ч п)
-- 16:00 - 18:00 (Гарри Кодер)
-- 20:00 — 22:10 (Властелин лаб...)  
+Кинозал 3  
+9. 12:00 - 13:30  
+   Люди в черных пикселях  
+10. 16:00 - 18:30  
+    Гарри Кодер и С++  
+11. 20:00 — 23:20  
+    Властелин лаб: Возвращение сеньора
