--- cinema/sessions_two_one.txt
@@ -1,28 +1,0 @@
-Cinema room 1
-4 
-10:00 - 12:00 
-Пятый алгоритм
-13:00 -16:15 
-Титаник: крах программы
-16:30 - 19:40 
-А дедлайны здесь жесткие
-20:00 - 23:20  
-Властелин лаб: Возвращение сеньора
-Cinema room 2
-3
-10:00 — 13:30 
-Шрек и болото багов
-16:00 - 18:00 
-Назад к компилятору
-20:00 - 22:15 
-Дебаггер 2: Судный день компиляции
-Cinema room 3
-4 
-11:00 - 13:15 
-Матрица данных
-14:00 - 16:00  
-Унесенные сессией
-17:00 - 19:30 
-Гарри Кодер
-20:00 - 22:15 
-Лабные войны
